app-ai-base/pyproject.toml
[project]
name = "app-ai-base"
version = "1.0.0"
description = "LangChain ReAct AI agent powered by Claude — production-ready base"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.135",
    "uvicorn[standard]>=0.41",
    "python-multipart>=0.0.22",

    # Validation & config
    "pydantic>=2.12",
    "pydantic-settings>=2.13",

    # LangChain
    "langchain-core>=1.2",
    "langchain-anthropic>=1.3",
    "langchain-text-splitters>=1.1",
    "langchain-tavily>=0.2",

    # Vector DB (choose one via .env — see MILVUS_URI)
    "pymilvus>=2.6",
    "milvus-lite>=2.5",   # optional: Milvus Lite (no Docker, file-based)

    # Embeddings — BGE-M3 local model (free, multilingual, ~1 GB)
    "FlagEmbedding>=1.3",
    "modelscope>=1.34",
    "sentence-transformers>=5.2",
    "torch>=2.10",
    "transformers==4.44.2",
    "accelerate>=1.13",

    # Document parsing
    "pymupdf>=1.27",
    "pytesseract",
    "pillow",
    "python-docx",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
